webxr/depth-sensing/depth_sensing_staleView.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/webxr_util.js"></script>
<script src="../resources/webxr_test_asserts.js"></script>
<script src="../resources/webxr_test_constants.js"></script>
<script src="../resources/webxr_test_constants_fake_depth.js"></script>

<script>

const fakeDeviceInitParams = {
  supportedModes: ["immersive-ar"],
  views: VALID_VIEWS,
  supportedFeatures: ALL_FEATURES,
};

const testFunctionGenerator = function(isCpuOptimized) {
  return (session, controller, t, sessionObjects) => {
    let done = false;

    const views = new Set();

    return session.requestReferenceSpace('viewer').then((viewerSpace) => {
      const glBinding = new XRWebGLBinding(session, sessionObjects.gl);

      const secondRafCb = function(time, frame) {
        for(const view of views) {
          t.step(() => {
            assert_throws_dom("InvalidStateError",
                                () => isCpuOptimized ? frame.getDepthInformation(view)
                                                     : glBinding.getDepthInformation(view),
                                "getDepthInformation() should throw when run with stale XRView");
          });
        }

        done = true;
      };

      const firstRafCb = function(time, frame) {
        const pose = frame.getViewerPose(viewerSpace);
        for(const view of pose.views) {
          views.add(view);
        }

        session.requestAnimationFrame(secondRafCb);
      };

      session.requestAnimationFrame(firstRafCb);

      return t.step_wait(() => done);
    });
  };
};

xr_session_promise_test("Ensures getDepthInformation() throws when run with stale XRView, `cpu-optimized`",
  testFunctionGenerator(/*isCpuOptimized=*/true),
  fakeDeviceInitParams,
  'immersive-ar', {
    requiredFeatures: ['depth-sensing'],
    depthSensing: {
      usagePreference: ['cpu-optimized'],
      dataFormatPreference: ['luminance-alpha', 'float32'],
    }
  },
  /*properties=*/null, /*glcontextPropertiesParam=*/null, /*gllayerPropertiesParam=*/null,
  /*ignoreSessionCreationFailure=*/true);

xr_session_promise_test("Ensures getDepthInformation() throws when not run with stale XRView, `gpu-optimized`",
  testFunctionGenerator(/*isCpuOptimized=*/false),
  fakeDeviceInitParams,
  'immersive-ar', {
    requiredFeatures: ['depth-sensing'],
    depthSensing: {
      usagePreference: ['gpu-optimized'],
      dataFormatPreference: ['luminance-alpha', 'float32'],
    }
  },
  /*properties=*/null, /*glcontextPropertiesParam=*/null, /*gllayerPropertiesParam=*/null,
  /*ignoreSessionCreationFailure=*/true);

</script>
